app/pyproject.toml
[tool.poetry]
name = "soltran"
version = "0.1.0"
description = ""
authors = ["Anton Pomieshchenko <ferumflex@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<=3.13"
pydantic = "^2.7.4"
pydantic-settings = "^2.3.3"
langchain-core = "^0.2.7"
langchain = "^0.2.5"
langsmith = "^0.1.77"
python-telegram-bot = {extras = ["all"], version = "^21.3"}
langchain-community = "^0.2.5"
openai = "^1.34.0"
solana = "^0.34.2"
langchain-openai = "^0.1.8"
langchainhub = "^0.1.20"
crewai = {extras = ["tools"], version = "^0.67.1"}


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
